// src/modals/types.ts
export enum ModalTypes {
  OrderDetails = 'OrderDetails',
  ConfirmPaying = 'ConfirmPaying',
  ConfirmRefunding = 'ConfirmRefunding',
  CancelCompleteOrder = 'CancelCompleteOrder',
  ProductDetails = 'ProductDetails',
  ProductCreate = 'ProductCreate',
  ProductAddonCreate = 'ProductAddonCreate',
  ProductAddonDetails = 'ProductAddonDetails',
  CashbookCreate = 'CashbookCreate',
  CashbookDetails = 'CashbookDetails',
  DebtCreate = 'DebtCreate',
  DebtCreatePayment = 'DebtCreatePayment',
  DebtDetails = 'DebtDetails',
  ContactCreate = 'ContactCreate',
  ContactUpdate = 'ContactUpdate',
  ContactDetails = 'ContactDetails',
  ContactGroupCreate = 'ContactGroupCreate',
  ContactGroupUpdate = 'ContactGroupUpdate',
  ContactGroupDetails = 'ContactGroupDetails',
  FrequentlyQuestionCreate = 'FrequentlyQuestionCreate',
  FrequentlyQuestionDetails = 'FrequentlyQuestionDetails',
  ReportMessage = 'ReportMessage',
  LabelMessageCreate = 'LabelMessageCreate',
  LabelMessageDetails = 'LabelMessageDetails',
  QuickMessageCreate = 'QuickMessageCreate',
  QuickMessageDetails = 'QuickMessageDetails',
  AbsenceTimeSelectDetails = 'AbsenceTimeSelectDetails',
  ChatbotPreview = 'ChatbotPreview',
  HistoryOrderTable = 'HistoryOrderTable',
  IngredientCreate = 'IngredientCreate',
  IngredientDetails = 'IngredientDetails',
  RecipeUpdate = 'RecipeUpdate',
  RecipeCreate = 'RecipeCreate',
  ContactLabels = 'ContactLabels',
  CreateImportGoods = 'CreateImportGoods',
  ImportGoodsDetails = 'ImportGoodsDetails',
  ExportGoodsDetails = 'ExportGoodsDetails',
  CreateExportGoods = 'CreateExportGoods',
  AutoLabels = 'AutoLabels',
  WarehouseDetails = 'WarehouseDetails',
  CancelPurchaseOrder = 'CancelPurchaseOrder',
  DebitPaymentPurchaseOrder = 'DebitPaymentPurchaseOrder',
  CreateStockTaking = 'CreateStockTaking',
  DetailStockTaking = 'DetailStockTaking',
}

export enum ModalPlacement {
  Left = 'left',
  Right = 'right',
  Top = 'top',
  Bottom = 'bottom',
}

export enum ModalSize {
  Full = 'full',
  Large = 'lg',
  Medium = 'md',
  Small = 'sm',
  Xsmall = 'xs',
}